Convert timings.browser NR-138521
